Add some error handling
In case the user quits dmenu by pressing <esc> the return code will be
non zero. This case is handled now. Also if the user types something
into the dmenu field, that is not a valid choise, that case is handled
now.

Well handled is not quite correct. For now an error message is printed.