Add a return value
In order to be able to determine in the calling script whether or not
dmenu run successfully and with a valid choise, I added the return
value. This might be usefull if in the calling script you want to do
something more, in case the dmenu wrapper was successful (or not).