Move scripts into src folder
I did that to have a tidy project direcotry tree. I will add a README.md
and LICENCE file later on (and maybe a Makefile) and those files are
supposed to be the only ones in the top level directory. Everything else
(like the scripts and e.g. some further documenatation, that I might add
in the future) should be in appropriate sub-direcotories.