Use slock's cmd option to suspend session
So sometimes after resuming from sleep, I had the problem that I
couldn't unlock the screen. I'm not really sure, but I could imagine
that this might have to do with the fact that earlier slock would have
been called after 'systemclt suspend' was called, but before the system
actually entered the sleep state. Maybe this messed something up. I'll
try it with this new version of the script and hope it helps. If now
I'll have to dig deeper.